# server/docs/swagger.yaml
basePath: /product-ms/v1
definitions:
  data.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      err_msg:
        type: string
    type: object
  data.CartItemBasicVO:
    properties:
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
      selected:
        type: boolean
      user_id:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  data.CartItemDetailVO:
    properties:
      id:
        type: integer
      product_info:
        $ref: '#/definitions/types.ProductSimplifiedInfo'
      quantity:
        type: integer
      selected:
        type: boolean
      status:
        description: '1: normal, 2: out of stock'
        type: integer
      total_price:
        type: integer
    type: object
  data.CartListVO:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/data.CartItemDetailVO'
        type: array
      selected_item_count:
        type: integer
      selected_price:
        type: integer
    type: object
  data.CartPriceEstimateResult:
    properties:
      product_price:
        type: integer
      shipping_price:
        type: integer
      tax:
        type: integer
      total:
        type: integer
    type: object
  data.ImgUploadRequest:
    properties:
      image_type:
        enum:
        - jpg
        - png
        - jpeg
        type: string
    required:
    - image_type
    type: object
  data.ImgUploadResponse:
    properties:
      image_id:
        type: string
      upload_url:
        type: string
    type: object
  types.ProductInfo:
    properties:
      capacity:
        type: string
      care_instructions:
        type: string
      category:
        type: string
      desc:
        type: string
      dimensions:
        type: string
      material:
        type: string
      name:
        type: string
      pic_info:
        type: string
      price:
        type: integer
      status:
        description: '0: 未上架, 1: 已上架'
        type: integer
      stock:
        type: integer
      weight:
        type: string
    type: object
  types.ProductSimplifiedInfo:
    properties:
      category:
        type: string
      desc:
        type: string
      id:
        type: integer
      name:
        type: string
      pic_info:
        type: string
      price:
        type: integer
      status:
        description: '0: 未上架, 1: 已上架'
        type: integer
      stock:
        type: integer
    type: object
  types.UpdateProductInfoRequest:
    properties:
      capacity:
        type: string
      care_instructions:
        type: string
      category:
        type: string
      desc:
        type: string
      dimensions:
        type: string
      id:
        type: integer
      material:
        type: string
      name:
        type: string
      pic_info:
        type: string
      price:
        type: integer
      weight:
        type: string
    type: object
  types.UpdateProductStatusRequest:
    properties:
      status:
        description: 0-新的状态是下架，1-新的状态是上架
        type: integer
    type: object
  types.UpdateProductStockRequest:
    properties:
      stock:
        type: integer
    type: object
info:
  contact: {}
  description: 商品微服务相关接口
  title: 商品服务 API
  version: "1.0"
paths:
  /customer/cart:
    get:
      consumes:
      - application/json
      description: Get user's cart info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/data.CartListVO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Get user's cart info
      tags:
      - Cart
  /customer/cart/items:
    post:
      consumes:
      - application/json
      description: Create a cart item
      parameters:
      - description: cart item info
        in: body
        name: cart_item
        required: true
        schema:
          $ref: '#/definitions/data.CartItemBasicVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/data.CartItemBasicVO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Create a cart item
      tags:
      - Cart
  /customer/cart/items/:item_id:
    delete:
      consumes:
      - application/json
      description: Delete a cart item
      parameters:
      - description: cart item ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Delete a cart item
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update a cart item
      parameters:
      - description: cart item info
        in: body
        name: cart_item
        required: true
        schema:
          $ref: '#/definitions/data.CartItemBasicVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/data.CartItemBasicVO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Update a cart item
      tags:
      - Cart
  /customer/cart/price-estimate:
    get:
      consumes:
      - application/json
      description: Calculate order price
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/data.CartPriceEstimateResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Calculate order price
      tags:
      - Cart
  /customer/cart/selected-num:
    get:
      consumes:
      - application/json
      description: Get number of selected items in cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: integer
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Get number of selected items in cart
      tags:
      - Cart
  /customer/product/{id}:
    get:
      consumes:
      - application/json
      description: 根据商品ID获取商品详细信息
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.ProductInfo'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "404":
          description: 商品不存在
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: 获取商品详情(用户侧)
      tags:
      - 商品
  /customer/products:
    get:
      consumes:
      - application/json
      description: 支持按关键词搜索、分类筛选、分页，并按更新时间排序
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 商品分类
        in: query
        name: category
        type: string
      - description: 偏移量，默认0
        in: query
        name: offset
        type: integer
      - description: 排序方式：0-按更新时间降序，1-按更新时间升序，默认0
        in: query
        name: order_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: 用户端获取商品列表
      tags:
      - 商品
  /merchant/images/upload-urls:
    post:
      consumes:
      - application/json
      description: Get presigned URL for image upload
      parameters:
      - description: image_type=(jpg|jpeg|png)
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/data.ImgUploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.ImgUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Get presigned URL for image upload
      tags:
      - Image
  /merchant/product/{id}:
    get:
      consumes:
      - application/json
      description: 根据商品ID获取商品详细信息
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.ProductInfo'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "404":
          description: 商品不存在
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: 获取商品详情
      tags:
      - 商品
  /merchant/products:
    get:
      consumes:
      - application/json
      description: 支持按关键词搜索、分类筛选、分页，并按更新时间排序
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 商品分类
        in: query
        name: category
        type: string
      - description: 偏移量，默认0
        in: query
        name: offset
        type: integer
      - description: 排序方式：0-按更新时间降序，1-按更新时间升序，默认0
        in: query
        name: order_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: 商家端获取商品列表
      tags:
      - 商品
    post:
      consumes:
      - application/json
      description: 新增一个商品
      parameters:
      - description: 商品信息
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/types.ProductInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: 添加商品
      tags:
      - 商品
  /merchant/products/:id/review/:decision:
    post:
      consumes:
      - application/json
      description: 审核拒绝或者通过
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: string
      - description: 审核状态
        enum:
        - approved
        - rejected
        in: path
        name: decision
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "404":
          description: 商品不存在
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: 审核商品
      tags:
      - 商品
  /merchant/products/:id/status:
    patch:
      consumes:
      - application/json
      description: 将商品状态更改为审核中或者下架
      parameters:
      - description: 商品状态修改请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.UpdateProductStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "404":
          description: 商品不存在
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: 商品审核或下架
      tags:
      - 商品
  /merchant/products/:id/stock:
    patch:
      consumes:
      - application/json
      description: 只有当商品处于下架状态时，才能更改商品库存
      parameters:
      - description: 更新商品库存请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.UpdateProductStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: 商家端更新商品库存
      tags:
      - 商品
  /merchant/products/{id}:
    put:
      consumes:
      - application/json
      description: 根据商品ID更新商品详细信息
      parameters:
      - description: 编辑商品请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.UpdateProductInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 编辑成功
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "404":
          description: 商品不存在
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: 编辑商品信息
      tags:
      - 商品
swagger: "2.0"